requirements.txt
bokeh==2.4.2
django-extensions==3.1.3
django-htmx==1.6.0